Done delete and update. Now need check if id available or not then can delete.

+++ app/src/main/java/com/example/a16031940/sqlitetester/ViewMarks.java
@@ -1,15 +1,29 @@
 package com.example.a16031940.sqlitetester;
 
+import android.content.Context;
+import android.content.Intent;
 import android.database.Cursor;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.ContextMenu;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.ImageView;
 import android.widget.ListView;
+import android.widget.Switch;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import java.lang.reflect.Array;
 import java.util.ArrayList;
+import java.util.List;
+
+import io.ghyeok.stickyswitch.widget.StickySwitch;
 
 public class ViewMarks extends AppCompatActivity {
 
@@ -17,19 +31,18 @@ public class ViewMarks extends AppCompatActivity {
     ArrayAdapter AA;
     ArrayList<Marks> marks;
     DatabaseHelper myDb;
+    Marks marker;
+    Switch switcher;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_marks);
         lv = findViewById(R.id.lv);
-
-
         marks = new ArrayList<Marks>();
         myDb = new DatabaseHelper(this);
-
         Cursor data = myDb.getAllData();
-
+        switcher = findViewById(R.id.switch1);
         if (data.getCount() == 0) {
             Toast.makeText(ViewMarks.this, "Nothing to view :(", Toast.LENGTH_LONG).show();
         } else {
@@ -41,19 +54,29 @@ public class ViewMarks extends AppCompatActivity {
 
         }
 
+
         lv.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
 
-//                Toast.makeText(getBaseContext(),marks.get(position).getName() + " ", Toast.LENGTH_SHORT).show();
-                Integer deletedRows = myDb.deleteData(marks.get(position).getId());
+                int[] info = {position,0};
 
-                if (deletedRows > 0) {
-                    Toast.makeText(ViewMarks.this, "Data is deleted", Toast.LENGTH_SHORT).show();
+
+                if (switcher.isChecked() == false) {
+
+
+                    Integer deletedRows = myDb.deleteData(marks.get(position).getId());
+
+                    if (deletedRows > 0) {
+                        Toast.makeText(ViewMarks.this, "Data is deleted", Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(ViewMarks.this, "Data is not deleted please try again", Toast.LENGTH_SHORT).show();
+                    }
                 } else {
-                    Toast.makeText(ViewMarks.this, "Data is not deleted please try again", Toast.LENGTH_SHORT).show();
+                    Intent i = new Intent(ViewMarks.this, Pop.class);
+                    i.putExtra("info",info);
+                    startActivity(i);
                 }
-
             }
         });
